script/pyproject.toml
[project]
name = "licensator"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Lorenzo Carbonell <a.k.a. atareao>", email = "lorenzo.carbonell.cerezo@gmail.com" }
]
dependencies = [
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/licensator"]
